Update r.hydro.flatten.py to meet 3DEP shoreline requirements
The USGS 3DEP standards require that the shoreline of a water body must be above the level of the modeled impoundment surface  in a DEM. This edit compares the elevation of the modeled water surface to the shoreline elevations. For estimated elevations of the shoreline that are less than or equal to the modeled water surface, the elevation of the shoreline is raised to .001 vertical units) above the modeled water level.   Thee addition of .001 meters or .001 feet is  well below the .1 meter  ( 10cm) vertical accuracy of the current standards for USGS QL2 and QL1 LiDAR collection.